Move sandbox code to node-core
