+++ packages/node-core/package.json
@@ -30,6 +30,7 @@
     "lodash": "^4.17.21",
     "prom-client": "^14.0.1",
     "sequelize": "6.23.0",
+    "vm2": "^3.9.9",
     "yargs": "^16.2.0"
   }
 }
