+++ packages/node-core/src/indexer/index.ts
@@ -11,3 +11,4 @@ export * from './poi.service';
 export * from './mmr.service';
 export * from './worker';
 export * from './dictionary.service';
+export * from './sandbox';
