+++ packages/node-core/test/sandbox/main.js
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 'use strict';
-Object.defineProperty(exports, '__esModule', { value: true });
+Object.defineProperty(exports, '__esModule', {value: true});
 exports.testSandbox = void 0;
 async function testSandbox() {
   await new Promise((resolve) => setTimeout(resolve, 3000));
