+++ packages/node/src/indexer/ds-processor.service.ts
@@ -12,7 +12,7 @@ import {
   SubstrateDatasourceProcessor,
   SubstrateNetworkFilter,
 } from '@subql/common-substrate';
-import { getLogger, NodeConfig } from '@subql/node-core';
+import { getLogger, NodeConfig, Sandbox } from '@subql/node-core';
 import {
   SecondLayerHandlerProcessor_0_0_0,
   SecondLayerHandlerProcessor_1_0_0,
@@ -22,7 +22,6 @@ import {
 
 import { VMScript } from 'vm2';
 import { SubqueryProject } from '../configure/SubqueryProject';
-import { Sandbox } from './sandbox.service';
 
 export interface DsPluginSandboxOption {
   root: string;
